fix(canon): keep virtual_project/materialize.rs under the source-length budget
CI's source-file-length gate (350 lines) rejected the growth to 359. Fold package node_modules mirror computation, link creation, and preserved-root assembly into a single materialize_package_node_modules call, and move write_declaration_tsconfig next to the helpers it drives in declaration_emit.rs. No behavior change; declaration-filtered test failures match the pre-change set on this machine exactly.
