+++ crates/vize_canon/src/batch/virtual_project/package_node_modules.rs
@@ -62,35 +62,39 @@ fn package_node_modules_mirrors(
     mirrors
 }
 
+/// (Re)create the mirror symlinks. Best-effort like the runtime-deps links:
+/// on a platform without symlink permission the run keeps the pre-fix
+/// resolution behavior instead of failing materialization.
+fn materialize_mirrors(mirrors: &[(PathBuf, PathBuf)]) {
+    for (target, source) in mirrors {
+        let source = vize_carton::path::canonicalize_non_verbatim(source);
+        if symlink_path(&source, target).is_err() {
+            continue;
+        }
+    }
+}
+
 impl VirtualProject {
-    /// Real per-package `node_modules` directories the mirror must expose at
-    /// the matching project-relative location, computed from every registered
-    /// original path (checked sources and passthrough dependencies alike).
-    pub(super) fn package_node_modules_mirrors(&self) -> Vec<(PathBuf, PathBuf)> {
-        package_node_modules_mirrors(
+    /// Mirror per-package `node_modules` links into the virtual tree and
+    /// return the roots the materialize GC must preserve: the special-cased
+    /// runtime-deps mirror plus every per-package link target computed from
+    /// the registered original paths (checked sources and passthrough
+    /// dependencies alike). Folded into one call so `materialize` stays under
+    /// the source-length budget.
+    pub(super) fn materialize_package_node_modules(&self) -> CorsaResult<Vec<PathBuf>> {
+        let mirrors = package_node_modules_mirrors(
             &self.project_root,
             &self.virtual_root,
             self.original_index
                 .keys()
                 .cloned()
                 .chain(self.passthrough_files.values().cloned()),
-        )
-    }
-
-    /// (Re)create the mirror symlinks. Best-effort like the runtime-deps
-    /// links: on a platform without symlink permission the run keeps the
-    /// pre-fix resolution behavior instead of failing materialization.
-    pub(super) fn materialize_package_node_modules(
-        &self,
-        mirrors: &[(PathBuf, PathBuf)],
-    ) -> CorsaResult<()> {
-        for (target, source) in mirrors {
-            let source = vize_carton::path::canonicalize_non_verbatim(source);
-            if symlink_path(&source, target).is_err() {
-                continue;
-            }
-        }
-        Ok(())
+        );
+        let mut preserved = Vec::with_capacity(mirrors.len() + 1);
+        preserved.push(self.virtual_root.join("node_modules"));
+        preserved.extend(mirrors.iter().map(|(target, _)| target.clone()));
+        materialize_mirrors(&mirrors);
+        Ok(preserved)
     }
 }
